# openapi-ad-db.yaml
openapi: 3.0.3
info:
  title: Lead Management API
  description: |
    광고 리드(DB) 데이터를 관리하는 REST API입니다.
    이벤트/사이트/광고주/매체별 리드를 등록, 조회, 수정, 삭제할 수 있습니다.

    ## 인증
    모든 API 요청에는 `x-api-key` 헤더가 필요합니다.

    ## 중복 체크
    동일한 `phone` + `event` 조합은 중복으로 처리되어 등록이 거부됩니다.
  version: 1.2.0
  contact:
    name: API Support

servers:
  - url: https://tst-db.vercel.app
    description: Production

security:
  - ApiKeyAuth: []

tags:
  - name: leads
    description: 리드(DB) 관리

paths:
  /api/leads:
    post:
      tags:
        - leads
      summary: 리드 등록
      description: |
        새로운 리드(DB) 데이터를 등록합니다.
        동일한 phone + event 조합이 이미 존재하면 409 에러를 반환합니다.
        IP 주소는 서버에서 자동으로 기록됩니다.
      operationId: createLead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadCreateRequest"
            example:
              event: "177"
              site: "26020301"
              advertiser: "이심당치과의원"
              media: "페이스북"
              category: "임플란트 가격노출"
              name: "홍길동"
              phone: "010-1234-5678"
              age: 59
              gender: "남"
              address: "서울시 강남구"
              email: "hong@example.com"
              survey1: "예"
              status: "대기"
              memo: "메모 내용"
      responses:
        "201":
          description: 리드 등록 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadResponse"
              example:
                code: 201
                message: "리드가 등록되었습니다."
                data:
                  id: 1
                  event: "177"
                  site: "26020301"
                  advertiser: "이심당치과의원"
                  media: "페이스북"
                  category: "임플란트 가격노출"
                  name: "홍길동"
                  phone: "010-1234-5678"
                  age: 59
                  gender: "남"
                  branch: null
                  address: "서울시 강남구"
                  email: "hong@example.com"
                  survey1: "예"
                  survey2: null
                  survey3: null
                  survey4: null
                  survey5: null
                  survey6: null
                  status: "대기"
                  memo: "메모 내용"
                  ip: "121.134.155.129"
                  createdAt: "2026-02-12T11:15:13.755Z"
        "400":
          description: 잘못된 요청 (필수 필드 누락 또는 유효하지 않은 데이터)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
              example:
                code: 400
                message: "필수 필드가 누락되었습니다."
                errors:
                  - field: name
                    message: "이름은 필수입니다."
                  - field: phone
                    message: "연락처 형식이 올바르지 않습니다. (예: 010-1234-5678)"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: 중복 데이터 (동일 phone + event 조합)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 409
                message: "이미 등록된 연락처입니다."

    get:
      tags:
        - leads
      summary: 리드 목록 조회
      description: |
        등록된 리드 목록을 조회합니다.
        필터링 및 페이지네이션을 지원합니다.
        결과는 id 기준 오름차순 정렬됩니다.
      operationId: listLeads
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - name: event
          in: query
          description: 이벤트 코드로 필터링
          schema:
            type: string
          example: "177"
        - name: site
          in: query
          description: 사이트 코드로 필터링
          schema:
            type: string
          example: "26020301"
        - name: advertiser
          in: query
          description: 광고주명으로 필터링
          schema:
            type: string
        - name: media
          in: query
          description: 매체로 필터링
          schema:
            type: string
        - name: status
          in: query
          description: 상태로 필터링
          schema:
            type: string
            enum:
              - 인정
              - 미인정
              - 대기
        - name: startDate
          in: query
          description: 조회 시작일 (등록일시 기준, 해당 날짜 포함)
          schema:
            type: string
            format: date
          example: "2026-02-08"
        - name: endDate
          in: query
          description: 조회 종료일 (등록일시 기준, 해당 날짜 포함)
          schema:
            type: string
            format: date
          example: "2026-02-12"
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadListResponse"
              example:
                code: 200
                message: "리드 목록을 조회했습니다."
                data:
                  content:
                    - id: 1
                      event: "177"
                      site: "26020301"
                      advertiser: "이심당치과의원"
                      media: "페이스북"
                      category: "임플란트 가격노출"
                      name: "홍길동"
                      phone: "010-1234-5678"
                      age: 59
                      gender: "남"
                      branch: null
                      address: "서울시 강남구"
                      email: "hong@example.com"
                      survey1: "예"
                      survey2: null
                      survey3: null
                      survey4: null
                      survey5: null
                      survey6: null
                      status: "대기"
                      memo: "메모 내용"
                      ip: "121.134.155.129"
                      createdAt: "2026-02-12T11:15:13.755Z"
                  page: 0
                  size: 20
                  totalElements: 1
                  totalPages: 1
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/leads/{leadId}:
    get:
      tags:
        - leads
      summary: 리드 상세 조회
      description: 특정 리드의 상세 정보를 조회합니다.
      operationId: getLead
      parameters:
        - $ref: "#/components/parameters/LeadIdParam"
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadResponse"
              example:
                code: 200
                message: "리드를 조회했습니다."
                data:
                  id: 1
                  event: "177"
                  site: "26020301"
                  advertiser: "이심당치과의원"
                  media: "페이스북"
                  category: "임플란트 가격노출"
                  name: "홍길동"
                  phone: "010-1234-5678"
                  age: 59
                  gender: "남"
                  branch: null
                  address: "서울시 강남구"
                  email: "hong@example.com"
                  survey1: "예"
                  survey2: null
                  survey3: null
                  survey4: null
                  survey5: null
                  survey6: null
                  status: "대기"
                  memo: "메모 내용"
                  ip: "121.134.155.129"
                  createdAt: "2026-02-12T11:15:13.755Z"
        "400":
          description: 유효하지 않은 ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 400
                message: "유효하지 않은 ID입니다."
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags:
        - leads
      summary: 리드 수정
      description: |
        특정 리드의 정보를 부분 수정합니다.
        변경할 필드만 전송하면 됩니다.
        빈 객체를 전송하면 400 에러를 반환합니다.
      operationId: updateLead
      parameters:
        - $ref: "#/components/parameters/LeadIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdateRequest"
            example:
              status: "미인정"
              memo: "연락 불가"
      responses:
        "200":
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadResponse"
              example:
                code: 200
                message: "리드가 수정되었습니다."
                data:
                  id: 1
                  event: "177"
                  site: "26020301"
                  advertiser: "이심당치과의원"
                  media: "페이스북"
                  category: "임플란트 가격노출"
                  name: "홍길동"
                  phone: "010-1234-5678"
                  age: 59
                  gender: "남"
                  branch: null
                  address: "서울시 강남구"
                  email: "hong@example.com"
                  survey1: "예"
                  survey2: null
                  survey3: null
                  survey4: null
                  survey5: null
                  survey6: null
                  status: "미인정"
                  memo: "연락 불가"
                  ip: "121.134.155.129"
                  createdAt: "2026-02-12T11:15:13.755Z"
        "400":
          description: 유효하지 않은 데이터 또는 변경할 필드 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
              examples:
                empty_body:
                  summary: 빈 요청
                  value:
                    code: 400
                    message: "변경할 필드가 없습니다."
                invalid_data:
                  summary: 유효하지 않은 데이터
                  value:
                    code: 400
                    message: "유효하지 않은 데이터입니다."
                    errors:
                      - field: phone
                        message: "연락처 형식이 올바르지 않습니다."
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - leads
      summary: 리드 삭제
      description: 특정 리드를 삭제합니다.
      operationId: deleteLead
      parameters:
        - $ref: "#/components/parameters/LeadIdParam"
      responses:
        "204":
          description: 삭제 성공 (응답 본문 없음)
        "400":
          description: 유효하지 않은 ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 400
                message: "유효하지 않은 ID입니다."
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API 인증 키

  parameters:
    LeadIdParam:
      name: leadId
      in: path
      required: true
      description: 리드 고유 번호
      schema:
        type: integer
        format: int64
      example: 1

    PageParam:
      name: page
      in: query
      description: 페이지 번호 (0부터 시작)
      schema:
        type: integer
        minimum: 0
        default: 0

    SizeParam:
      name: size
      in: query
      description: 페이지당 항목 수 (최대 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  responses:
    Unauthorized:
      description: 인증 실패 (API Key 누락 또는 불일치)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 401
            message: "인증이 필요합니다."

    NotFound:
      description: 리드를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 404
            message: "리드를 찾을 수 없습니다."

  schemas:
    LeadCreateRequest:
      type: object
      required:
        - event
        - site
        - advertiser
        - media
        - category
        - name
        - phone
      properties:
        event:
          type: string
          description: 이벤트 코드
          example: "177"
        site:
          type: string
          description: 사이트 코드
          example: "26020301"
        advertiser:
          type: string
          description: 광고주명
          example: "이심당치과의원"
        media:
          type: string
          description: 매체
          example: "페이스북"
        category:
          type: string
          description: 구분 (캠페인 유형)
          example: "임플란트 가격노출"
        name:
          type: string
          description: 이름
          example: "홍길동"
        phone:
          type: string
          description: "연락처 (형식: 01X-XXXX-XXXX)"
          pattern: "^01[0-9]-[0-9]{3,4}-[0-9]{4}$"
          example: "010-1234-5678"
        age:
          type: integer
          description: 나이
          minimum: 1
          maximum: 150
          nullable: true
          example: 59
        gender:
          type: string
          description: 성별
          nullable: true
          enum:
            - 남
            - 여
          example: "남"
        branch:
          type: string
          description: 지점
          nullable: true
        address:
          type: string
          description: 주소
          nullable: true
          example: "서울시 강남구"
        email:
          type: string
          description: 이메일
          format: email
          nullable: true
          example: "hong@example.com"
        survey1:
          type: string
          description: 설문1
          nullable: true
        survey2:
          type: string
          description: 설문2
          nullable: true
        survey3:
          type: string
          description: 설문3
          nullable: true
        survey4:
          type: string
          description: 설문4
          nullable: true
        survey5:
          type: string
          description: 설문5
          nullable: true
        survey6:
          type: string
          description: 설문6
          nullable: true
        status:
          type: string
          description: 상태
          enum:
            - 인정
            - 미인정
            - 대기
          default: "대기"
          example: "대기"
        memo:
          type: string
          description: 메모
          nullable: true

    LeadUpdateRequest:
      type: object
      description: 변경할 필드만 포함합니다. 모든 필드는 선택사항이며, 빈 객체는 허용되지 않습니다.
      properties:
        event:
          type: string
          description: 이벤트 코드
        site:
          type: string
          description: 사이트 코드
        advertiser:
          type: string
          description: 광고주명
        media:
          type: string
          description: 매체
        category:
          type: string
          description: 구분 (캠페인 유형)
        name:
          type: string
          description: 이름
        phone:
          type: string
          description: "연락처 (형식: 01X-XXXX-XXXX)"
          pattern: "^01[0-9]-[0-9]{3,4}-[0-9]{4}$"
        age:
          type: integer
          description: 나이
          minimum: 1
          maximum: 150
          nullable: true
        gender:
          type: string
          description: 성별
          nullable: true
          enum:
            - 남
            - 여
        branch:
          type: string
          description: 지점
          nullable: true
        address:
          type: string
          description: 주소
          nullable: true
        email:
          type: string
          description: 이메일
          format: email
          nullable: true
        survey1:
          type: string
          description: 설문1
          nullable: true
        survey2:
          type: string
          description: 설문2
          nullable: true
        survey3:
          type: string
          description: 설문3
          nullable: true
        survey4:
          type: string
          description: 설문4
          nullable: true
        survey5:
          type: string
          description: 설문5
          nullable: true
        survey6:
          type: string
          description: 설문6
          nullable: true
        status:
          type: string
          description: 상태
          enum:
            - 인정
            - 미인정
            - 대기
        memo:
          type: string
          description: 메모
          nullable: true

    LeadData:
      type: object
      properties:
        id:
          type: integer
          description: 리드 고유 번호
          readOnly: true
          example: 1
        event:
          type: string
          description: 이벤트 코드
          example: "177"
        site:
          type: string
          description: 사이트 코드
          example: "26020301"
        advertiser:
          type: string
          description: 광고주명
          example: "이심당치과의원"
        media:
          type: string
          description: 매체
          example: "페이스북"
        category:
          type: string
          description: 구분 (캠페인 유형)
          example: "임플란트 가격노출"
        name:
          type: string
          description: 이름
          example: "홍길동"
        phone:
          type: string
          description: 연락처
          example: "010-1234-5678"
        age:
          type: integer
          description: 나이
          nullable: true
          example: 59
        gender:
          type: string
          description: 성별
          nullable: true
          example: "남"
        branch:
          type: string
          description: 지점
          nullable: true
          example: null
        address:
          type: string
          description: 주소
          nullable: true
          example: "서울시 강남구"
        email:
          type: string
          description: 이메일
          nullable: true
          example: "hong@example.com"
        survey1:
          type: string
          description: 설문1
          nullable: true
          example: null
        survey2:
          type: string
          description: 설문2
          nullable: true
          example: null
        survey3:
          type: string
          description: 설문3
          nullable: true
          example: null
        survey4:
          type: string
          description: 설문4
          nullable: true
          example: null
        survey5:
          type: string
          description: 설문5
          nullable: true
          example: null
        survey6:
          type: string
          description: 설문6
          nullable: true
          example: null
        status:
          type: string
          description: 상태
          example: "대기"
        memo:
          type: string
          description: 메모
          nullable: true
          example: null
        ip:
          type: string
          description: 등록 IP (서버에서 자동 기록)
          readOnly: true
          example: "121.134.155.129"
        createdAt:
          type: string
          description: 등록일시 (서버에서 자동 기록)
          format: date-time
          readOnly: true
          example: "2026-02-12T11:15:13.755Z"

    LeadResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          description: HTTP 상태 코드
          example: 200
        message:
          type: string
          description: 성공 메시지
          example: "리드를 조회했습니다."
        data:
          $ref: "#/components/schemas/LeadData"

    LeadListResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          description: HTTP 상태 코드
          example: 200
        message:
          type: string
          description: 성공 메시지
          example: "리드 목록을 조회했습니다."
        data:
          type: object
          required:
            - content
            - page
            - size
            - totalElements
            - totalPages
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/LeadData"
            page:
              type: integer
              description: 현재 페이지 번호
              example: 0
            size:
              type: integer
              description: 페이지당 항목 수
              example: 20
            totalElements:
              type: integer
              format: int64
              description: 전체 항목 수
              example: 14
            totalPages:
              type: integer
              description: 전체 페이지 수
              example: 1

    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: HTTP 상태 코드
          example: 400
        message:
          type: string
          description: 에러 메시지
          example: "잘못된 요청입니다."
        errors:
          type: array
          description: 필드별 상세 에러 목록
          items:
            type: object
            properties:
              field:
                type: string
                description: 에러 필드명
              message:
                type: string
                description: 필드별 에러 메시지

    ValidationErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"